Fixed issue with double value, where int was expected

+++ app/src/main/java/app/com/vshkl/veryweather/currentweather/weather/Snow.java
@@ -8,19 +8,19 @@ public class Snow {
 
     @SerializedName("3h")
     @Expose
-    private int _3h;
+    private double _3h;
 
     /**
      * @return The _3h
      */
-    public int get3h() {
+    public double get3h() {
         return _3h;
     }
 
     /**
      * @param _3h The 3h
      */
-    public void set3h(int _3h) {
+    public void set3h(double _3h) {
         this._3h = _3h;
     }
